style: update color and display properties in CSS
The color property for the `.dull` class has been updated from `var(--color-mono-80)` to `var(--color-mono-90)` to make the text slightly darker.

The display property for the `.spacer` class has been changed from `flex` to `block` and a new `height` property has been added with the value `var(--size-base)` to adjust the spacing.